Added support for embedded columns ---------------------------------- Wordpress will not allow for embedded short codes with the same name. For example: [one_half]   [one_half][/one_half] [/one_half]
Use of these shortcodes will cause the page to be formatted improperly
and will leave the closing [/one_half] text in the page.

By adding column names with "_inner" on the end that perform identically
to half and third (with lasts), Wordpress will execute the short codes
correctly.  For example:
[one_half]
  [one_half_inner][/one_half_inner]
[/one_half]

Half and Third were chosen as dividing the page further is likely to
create tiny, unusable columns.  Adding additional columns sizes is
trivial (modify line 416).